Move most of rendering logic to ShaderWrapper
